refactor: streamline stock data retrieval by integrating getStock function and removing redundant code
